back/requirements.txt
asgiref==3.5.0; python_version >= '3.7'

django-cors-headers==3.11.0

django-mongoengine==0.4.6

django==3.2.13

djangorestframework==3.13.1

djongo==1.3.6

dnspython==2.2.1

mongoengine==0.24.1

pymongo==3.12.1

pytz==2022.1

sqlparse==0.2.4; python_version >= '3.5'